# entity-service/openapi.yaml
openapi: 3.0.1
info:
  title: Entity Service API
  version: 1.0.0
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /change-requests:
    post:
      summary: Create a change request (ServiceNow data source only).
      operationId: createChangeRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChangeRequestRequest'
      responses:
        "201":
          description: Change request created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateChangeRequestResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /change-requests/{id}:
    patch:
      summary: Update a change request (ServiceNow data source only).
      operationId: patchChangeRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchChangeRequestRequest'
      responses:
        "200":
          description: The updated change request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatchChangeRequestResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Change request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get a single change request by ID (ServiceNow data source only).
      operationId: getChangeRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The change request detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequest'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Change request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /change-requests/{id}/approvals:
    get:
      summary: Get the approval stages for a change request (ServiceNow data source only).
      operationId: getChangeRequestApprovals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The change request's approval stages.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestApprovals'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Change request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /change-requests/{id}/approvals/decision:
    post:
      summary: Decide the caller's own pending approval on a change request.
      description: >-
        Records the caller's decision on their own pending approval stage. Only the
        caller's own pending approval can be acted on; the change request's own state
        cascades automatically in the backing data source.
      operationId: decideChangeRequestApproval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRequestApprovalDecisionRequest'
      responses:
        "200":
          description: The decided approval record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestApprovalDecisionResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Change request or pending approval not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /change-requests/search:
    post:
      summary: Search change requests (ServiceNow data source only).
      operationId: searchChangeRequests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchChangeRequestsRequest'
      responses:
        "200":
          description: Change requests matching the filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchChangeRequestsResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /change-requests/group-by:
    post:
      summary: >-
        Server-side aggregation of change requests by a single field
        (ServiceNow data source only).
      operationId: groupChangeRequestsBy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupChangeRequestsByRequest'
      responses:
        "200":
          description: Aggregated change request counts per value of groupBy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupByResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /incident-tasks/search:
    post:
      summary: Search incident tasks (ServiceNow data source only). Search and get only -- no create/update.
      operationId: searchIncidentTasks
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchIncidentTasksRequest'
      responses:
        "200":
          description: Paginated list of incident tasks matching the filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchIncidentTasksResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /incident-tasks/group-by:
    post:
      summary: >-
        Server-side aggregation of incident tasks by a single field
        (ServiceNow data source only).
      operationId: groupIncidentTasksBy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupIncidentTasksByRequest'
      responses:
        "200":
          description: Aggregated incident task counts per value of groupBy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupByResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /incident-tasks/{id}:
    get:
      summary: Get a single incident task by ID (ServiceNow data source only).
      operationId: getIncidentTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The incident task detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentTaskDetail'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Incident task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    GroupByResponse:
      type: object
      required: [groups, othersCount, totalRecords]
      description: >-
        The shared result shape for every resource's dedicated group-by
        aggregation endpoint.
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/GroupByBucket'
          description: The top buckets by count, capped at the request's maxGroups.
        othersCount:
          type: integer
          description: >-
            The number of matching records folded together because their
            bucket fell outside the top maxGroups.
        totalRecords:
          type: integer
          description: >-
            Total number of records matching the request's filters; always
            equals the sum of groups' counts plus othersCount.
...
    SearchChangeRequestsRequest:
      type: object
      properties:
        filters:
          type: object
          properties:
            projectIds:
              type: array
              items:
                type: string
                format: uuid
            searchQuery:
              type: string
            states:
              type: array
              items:
                type: string
                enum: [new, assess, authorize, customer_approval, scheduled, implement, review, customer_review, rollback, closed, canceled]
            impacts:
              type: array
              items:
                type: string
                enum: [high, medium, low]
            closedStartDate:
              type: string
              format: date-time
              description: "ISO-8601 UTC datetime (YYYY-MM-DDTHH:MM:SSZ)"
            closedEndDate:
              type: string
              format: date-time
              description: "ISO-8601 UTC datetime (YYYY-MM-DDTHH:MM:SSZ)"
            number:
              type: string
              description: >-
                A single change request number (e.g. "CHG0010001"); matches
                exactly. Routed as a first-class filter rather than through
                the free-text searchQuery scan.
            filters:
              type: array
              items:
                $ref: '#/components/schemas/ChangeRequestFieldFilter'
              description: >-
                A generic field/op/values filter array, additive to the named
                filters above. See ChangeRequestFieldFilter for the supported
                field/op combinations.
        sortBy:
          type: object
          required: [field, order]
          properties:
            field:
              type: string
              enum: [createdOn, updatedOn]
            order:
              type: string
              enum: [asc, desc]
        pagination:
          type: object
          properties:
            offset:
              type: integer
              minimum: 0
              default: 0
            limit:
              type: integer
              minimum: 1
              maximum: 50
              default: 20
    GroupChangeRequestsByRequest:
      type: object
      required: [groupBy]
      properties:
        filters:
          type: object
          description: Same shape as SearchChangeRequestsRequest.filters.
          properties:
            projectIds:
              type: array
              items:
                type: string
                format: uuid
            searchQuery:
              type: string
            states:
              type: array
              items:
                type: string
                enum: [new, assess, authorize, customer_approval, scheduled, implement, review, customer_review, rollback, closed, canceled]
            impacts:
              type: array
              items:
                type: string
                enum: [high, medium, low]
            closedStartDate:
              type: string
              format: date-time
              description: "ISO-8601 UTC datetime (YYYY-MM-DDTHH:MM:SSZ)"
            closedEndDate:
              type: string
              format: date-time
              description: "ISO-8601 UTC datetime (YYYY-MM-DDTHH:MM:SSZ)"
            number:
              type: string
              description: >-
                A single change request number (e.g. "CHG0010001"); matches
                exactly. Routed as a first-class filter rather than through
                the free-text searchQuery scan.
            filters:
              type: array
              items:
                $ref: '#/components/schemas/ChangeRequestFieldFilter'
              description: >-
                A generic field/op/values filter array, additive to the named
                filters above. See ChangeRequestFieldFilter for the supported
                field/op combinations.
        groupBy:
          type: string
          enum: [state, assignmentGroup]
          description: The field to aggregate change request counts by.
        maxGroups:
          type: integer
          description: >-
            Caps the response to the top maxGroups buckets by count; the
            remainder is folded into GroupByResponse.othersCount. Optional; a
            default is applied when omitted.
...
    SearchChangeRequestsResponse:
      type: object
      properties:
        changeRequests:
          type: array
          items:
            $ref: '#/components/schemas/SearchChangeRequestView'
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
    ChangeRequest:
      allOf:
        - $ref: '#/components/schemas/SearchChangeRequestView'
        - type: object
          properties:
            createdBy:
              type: string
            justification:
              type: string
              nullable: true
            impactDescription:
              type: string
              nullable: true
            serviceOutage:
              type: string
              nullable: true
            communicationPlan:
              type: string
              nullable: true
            rollbackPlan:
              type: string
              nullable: true
            testPlan:
              type: string
              nullable: true
            hasCustomerApproved:
              type: boolean
            hasCustomerReviewed:
              type: boolean
            approvedBy:
              $ref: '#/components/schemas/EntityRef'
              nullable: true
            approvedOn:
              type: string
              nullable: true
...
    ChangeRequestApprovals:
      type: object
      properties:
        approvals:
          type: array
          items:
            $ref: '#/components/schemas/ChangeRequestApproval'
    ChangeRequestApprovalDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, rejected]
          description: Caller's decision on their own pending approval.
    ChangeRequestApprovalDecisionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: UUID of the approval record that was decided.
        state:
          type: string
          description: Resulting state of the approval ("approved" or "rejected").
...
    CreateChangeRequestRequest:
      type: object
      required:
        - subject
      properties:
        subject:
          type: string
        category:
          type: string
          enum:
            - hardware
            - software
            - service
            - system_software
            - applications_software
            - network
            - telecom
            - documentation
            - other
            - regular_release_cloud
            - hotfix_release_cloud
            - devops
            - cloud_computing
        serviceId:
          type: string
          format: uuid
        serviceOfferingId:
          type: string
          format: uuid
        configurationItemId:
          type: string
          format: uuid
        priority:
          type: string
          enum:
            - critical
            - high
            - moderate
            - low
        impact:
          type: string
          enum:
            - high
            - medium
            - low
        type:
          type: string
          enum:
            - standard
            - normal
            - emergency
            - model
            - site_reliability_ops
            - azure
        state:
          type: string
          enum:
            - new
            - assess
            - authorize
            - customer_approval
            - scheduled
            - implement
            - review
            - customer_review
            - rollback
            - closed
            - canceled
        groupId:
          type: string
          format: uuid
        assignedEngineerId:
          type: string
          format: uuid
        risk:
          type: string
          enum:
            - high
            - moderate
            - low
        requestedById:
          type: string
          format: uuid
        description:
          type: string
        justification:
          type: string
        implementationPlan:
          type: string
        riskImpactAnalysis:
          type: string
        backoutPlan:
          type: string
        testPlan:
          type: string
        plannedStartDate:
          type: string
          description: Planned start date and time (format YYYY-MM-DD HH:MM:SS).
          pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$'
        plannedEndDate:
          type: string
          description: Planned end date and time (format YYYY-MM-DD HH:MM:SS).
          pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$'
        comment:
          type: string
        workNote:
          type: string
    CreateChangeRequestResponse:
      type: object
      properties:
        message:
          type: string
        changeRequest:
          type: object
          properties:
            id:
              type: string
              format: uuid
            number:
              type: string
            createdOn:
              type: string
            createdBy:
              type: string
    PatchChangeRequestRequest:
      type: object
      minProperties: 1
      description: >
        Any combination of fields may be supplied; at least one is required.
        Omitted fields are left unchanged.
      properties:
        title:
          type: string
        description:
          type: string
        projectId:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        deploymentId:
          type: string
          format: uuid
        deployedProductId:
          type: string
          format: uuid
        assignedEngineerId:
          type: string
          format: uuid
        assignedTeamId:
          type: string
          format: uuid
        plannedStartOn:
          type: string
          description: Planned start date and time (format YYYY-MM-DD HH:MM:SS).
          pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$'
        plannedEndOn:
          type: string
          description: Planned end date and time (format YYYY-MM-DD HH:MM:SS).
          pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$'
        impact:
          type: string
          enum:
            - high
            - medium
            - low
        state:
          type: string
          enum:
            - new
            - assess
            - authorize
            - customer_approval
            - scheduled
            - implement
            - review
            - customer_review
            - rollback
            - closed
            - canceled
        type:
          type: string
          enum:
            - standard
            - normal
            - emergency
            - model
            - site_reliability_ops
            - azure
        justification:
          type: string
        impactDescription:
          type: string
        serviceOutage:
          type: string
        communicationPlan:
          type: string
        rollbackPlan:
          type: string
        testPlan:
          type: string
        isCustomerApproved:
          type: boolean
        isCustomerReviewed:
          type: boolean
    PatchChangeRequestResponse:
      type: object
      required:
        - message
        - changeRequest
      properties:
        message:
          type: string
        changeRequest:
          $ref: '#/components/schemas/ChangeRequest'
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...
    SearchIncidentTasksRequest:
      type: object
      properties:
        filters:
          type: object
          properties:
            searchQuery:
              type: string
            number:
              type: string
              description: >-
                A single incident task number (e.g. "TASK0082453"); matches
                exactly. Routed as a first-class filter rather than through
                the free-text searchQuery scan.
            filters:
              type: array
              items:
                $ref: '#/components/schemas/IncidentTaskFieldFilter'
              description: >-
                A generic field/op/values filter array. See
                IncidentTaskFieldFilter for the supported field/op combinations.
        pagination:
          $ref: '#/components/schemas/Pagination'
    GroupIncidentTasksByRequest:
      type: object
      required: [groupBy]
      properties:
        filters:
          type: object
          description: Same shape as SearchIncidentTasksRequest.filters.
          properties:
            searchQuery:
              type: string
            number:
              type: string
              description: >-
                A single incident task number (e.g. "TASK0082453"); matches
                exactly. Routed as a first-class filter rather than through
                the free-text searchQuery scan.
            filters:
              type: array
              items:
                $ref: '#/components/schemas/IncidentTaskFieldFilter'
              description: >-
                A generic field/op/values filter array. See
                IncidentTaskFieldFilter for the supported field/op combinations.
        groupBy:
          type: string
          enum: [state, assignmentGroup]
          description: The field to aggregate incident task counts by.
        maxGroups:
          type: integer
          description: >-
            Caps the response to the top maxGroups buckets by count; the
            remainder is folded into GroupByResponse.othersCount. Optional; a
            default is applied when omitted.
...
    SearchIncidentTasksResponse:
      type: object
      properties:
        incidentTasks:
          type: array
          items:
            $ref: '#/components/schemas/IncidentTask'
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
    IncidentTaskDetail:
      type: object
      properties:
        id:
          type: string
          nullable: true
        number:
          type: string
          nullable: true
        subject:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
          description: Raw passthrough state value from the data source (see IncidentTaskFieldFilter).
        stateLabel:
          type: string
          nullable: true
        incident:
          $ref: '#/components/schemas/CaseNumberRef'
          nullable: true
        assignmentGroup:
          $ref: '#/components/schemas/EntityRef'
          nullable: true
        assignedTo:
          $ref: '#/components/schemas/EntityRef'
          nullable: true
        description:
          type: string
          nullable: true
        priority:
          type: string
          nullable: true
        openedOn:
          type: string
          nullable: true
        closedOn:
          type: string
          nullable: true
...